requirements3.txt
# $ virtualenv env -p /usr/bin/python3.6
# $ source env/bin/activate
# $ pip install -r requirements3.txt

Pillow==5.0.0
